refactor(desktop): move agent actions off the MCP path
The show-actions MCP tool never reached an agent: the PostHog MCP is
single-exec, so individual tools are unreachable. Drop the tool, its UI
app and the app-bridge method, and keep the plumbing a local tool needs.

The action types lose their `McpApp` prefix, the zod schema moves to
`core/agent-actions`, and the dispatch moves to the deep-link router,
which already resolves a scheme and calls `handleUrl`.
