+++ products/desktop/packages/core/src/mcp-apps/schemas.test.ts
@@ -1,7 +1,6 @@
 import { describe, expect, it } from "vitest";
 import {
   LEGACY_RESOURCE_URI_META_KEY,
-  mcpAppActionSchema,
   POSTHOG_EXEC_TOOL_KEY,
   resolveResultResourceUri,
 } from "./schemas";
@@ -45,52 +44,5 @@ describe("resolveResultResourceUri", () => {
 
   it("pins the built-in exec tool key", () => {
     expect(POSTHOG_EXEC_TOOL_KEY).toBe("mcp__posthog__exec");
-  });
-});
-
-describe("mcpAppActionSchema", () => {
-  // A card is sandboxed HTML any MCP server can supply, so a blank required field has to be
-  // rejected here. Letting it through leaves a button that silently does nothing when clicked.
-  it.each([
-    ["a blank compose prompt", { kind: "compose", prompt: "   " }],
-    ["a missing space id", { kind: "open_space", channel_id: "" }],
-    [
-      "a missing canvas id",
-      { kind: "open_canvas", channel_id: "chan", canvas_id: "" },
-    ],
-    [
-      "a blank canvas channel id",
-      { kind: "open_canvas", channel_id: "  ", canvas_id: "dash" },
-    ],
-    ["an unknown verb", { kind: "open_website", url: "https://evil.example" }],
-  ])("rejects %s", (_name, action) => {
-    expect(mcpAppActionSchema.safeParse(action).success).toBe(false);
-  });
-
-  it("drops a label rather than carrying it into a link", () => {
-    const parsed = mcpAppActionSchema.parse({
-      kind: "open_space",
-      channel_id: "chan",
-      label: "Open the space",
-    });
-
-    expect(parsed).toEqual({ kind: "open_space", channel_id: "chan" });
-  });
-
-  it("keeps an optional repo but drops a blank one", () => {
-    expect(
-      mcpAppActionSchema.parse({
-        kind: "compose",
-        prompt: "Do it",
-        repo: "posthog/posthog",
-      }),
-    ).toEqual({ kind: "compose", prompt: "Do it", repo: "posthog/posthog" });
-    expect(
-      mcpAppActionSchema.safeParse({
-        kind: "compose",
-        prompt: "Do it",
-        repo: "  ",
-      }).success,
-    ).toBe(false);
   });
 });
