+++ products/desktop/packages/core/src/mcp-apps/schemas.ts
@@ -149,32 +149,6 @@ export const openLinkInput = z.object({
   url: z.string(),
 });
 
-// Required fields are non-blank here rather than in the link builder, so a bad action fails at
-// the boundary instead of being rebuilt into a button that does nothing when clicked. `label` is
-// carried by the card and never reaches a link, so it is not declared.
-const requiredField = z.string().trim().min(1);
-
-export const mcpAppActionSchema = z.discriminatedUnion("kind", [
-  z.object({
-    kind: z.literal("compose"),
-    prompt: requiredField,
-    repo: z.string().trim().min(1).optional(),
-  }),
-  z.object({
-    kind: z.literal("open_space"),
-    channel_id: requiredField,
-  }),
-  z.object({
-    kind: z.literal("open_canvas"),
-    channel_id: requiredField,
-    canvas_id: requiredField,
-  }),
-]);
-
-export const openActionInput = z.object({
-  action: mcpAppActionSchema,
-});
-
 export const mcpAppsSubscriptionInput = z.object({
   toolKey: z.string(),
 });
