+++ products/desktop/packages/ui/src/features/mcp-apps/components/McpAppHost.tsx
@@ -107,7 +107,6 @@ export function McpAppHost({
     trpc.mcpApps.proxyResourceRead.mutationOptions(),
   );
   const openLinkMut = useMutation(trpc.mcpApps.openLink.mutationOptions());
-  const openActionMut = useMutation(trpc.mcpApps.openAction.mutationOptions());
 
   const { sendWhenReady } = useAppBridge({
     iframeEl,
@@ -132,7 +131,6 @@ export function McpAppHost({
       uri: string;
     }) => Promise<ReadResourceResult>,
     openLink: openLinkMut.mutateAsync,
-    openAction: openActionMut.mutateAsync,
   });
 
   const sentResultForCallRef = useRef<string | null>(null);
