+++ services/mcp/src/tools/index.ts
@@ -7,8 +7,6 @@ import parserRecipeCreate from './aiObservability/parserRecipeCreate'
 import parserRecipeReference from './aiObservability/parserRecipeReference'
 // Debug
 import debugMcpUiApps from './debug/debugMcpUiApps'
-// Desktop
-import taskShowActions from './desktop/taskShowActions'
 // Experiments (hand-written — CRUD + lifecycle are codegen in generated/experiments.ts)
 import getExperimentResults from './experiments/getResults'
 import experimentListDeprecated from './experiments/listDeprecated'
@@ -115,9 +113,6 @@ export const TOOL_MAP: Record<string, () => ToolBase<ZodObjectAny>> = {
     // Debug
     'debug-mcp-ui-apps': debugMcpUiApps,
     'loops-review': loopsReview,
-
-    // Desktop
-    'task-show-actions': taskShowActions,
 
     // Feedback
     'agent-feedback': submitFeedback,
